backend/requirements.txt
# Core Django dependencies
Django==5.1.5
djangorestframework==3.15.2
djangorestframework-simplejwt==5.4.0
django-cors-headers==4.6.0
python-dotenv==1.0.1

# PDF Processing
PyPDF2==3.0.1
pdfplumber==0.10.4

# AI/ML Integration
openai==1.57.0
scikit-learn>=1.3.0
numpy>=1.24.0

# Additional dependencies
requests==2.32.3
urllib3==2.2.3
